Fixed overzealous delete bug, where deleting (in the Ontology Editor Panel) one relationship from a term that had two relationships would cause the term to be silently deleted. Test case:  load so.obo, find insertion, show parents (in graph view).  Then in OEP, say "Delete" (which should delete the relationship that is shown but not "insertion" itself).  In buggy version, you could see the "insertion" node disappear in the graph view.  In fixed version, it moves over to its now single parent.
The problem was that OBOTermPanel.createSelectionFromPaths had been declared
as a static method, which is wrong.  As a consequence of that, the call inside
that method to SelectionManager.createSelectionFromPaths was unable to use "this"
as an argument and thus was called with null as the first argument; this also
resulted in the bad behavior.
The thing that bothers me is that I'm not sure why John did it that way, and
whether there might be any bad ramifications of my fix (though so far everything
seems to be ok).



git-svn-id: svn://svn.code.sf.net/p/geneontology/svn/java/oboedit/trunk@622 6f0e8829-b336-0410-acfb-cb9b228023ad