- Fixed bug #2023468 : if you had two OTEs in local mode, and you changed one of them to global, BOTH of them would show the current selection (the one that's still local mode shouldn't change).
Also, made local-select-mode bg color a very pale pinkish-beige (was light gray) because
Jen felt it didn't contrast enough with the pale-blue local selection color.
If I just make it a lighter gray, the difference between that and
white will become so subtle as to be unnoticeable, and if I make
the local-select selection color darker (as Jen suggested), it
will be too close to the global-select color.



git-svn-id: svn://svn.code.sf.net/p/geneontology/svn/java/oboedit/trunk@1416 6f0e8829-b336-0410-acfb-cb9b228023ad