Fixed problem where Synonym editor wouldn't close when you committed. Description of problem:
In the past (e.g., in OE-2.000), after you added a new synonym and clicked Commit, the synonym editor (with a place to add a Name and Xrefs) changed back to the synonym display, showing the synonym name with the Scope under it and (if the synonym was in red) the little fixer tool icon on the left. Now, in 2.1-b7, nothing visibly changes when you click commit--the synonym editor stays open rather than changing to the synonym display, so it looks like the edit hasn't "taken".

I discovered that this broke in v2.1-b4 because this line in TextEditor.commit was removed:
 			Preferences.getPreferences().fireReconfigEvent(new ReconfigEvent(this));
That made it look like Commit wasn't working, because, for example, if you were editing a synonym and you clicked Commit,
the Synonym Editor stayed open instead of closing and looking like a finished Synonym.
I wish there had been some sort of comment or commit message explaining why the line was removed.  I'm concerned that
maybe it was removed to fix some other problem, which will now reappear.



git-svn-id: svn://svn.code.sf.net/p/geneontology/svn/java/oboedit/trunk@3328 6f0e8829-b336-0410-acfb-cb9b228023ad